+++ CHANGELOG.md
@@ -3,6 +3,12 @@
 
 [comment]: <>  (2026-12-10 : YEAR-MONTH-DAY)
 
+## [1.6.0] - 2026-02-25
++ ADD Reload to the Read nodes
++ ADD Localization Policy to the Read nodes
++ FIX check/ uncheck tracks properties on Tracker nodes
++ ADD some missing docstrings
+
 ## [1.5.0] - 2026-02-25
 + ADD MotionBakery integration
 + ADD toggle T/R/S/ALL for tracks on Tracker nodes